+++ src/api/fetchLevels.ts
@@ -1,21 +1,5 @@
 import { fetchServerData } from './fetchServerData';
-
-export interface Level {
-  title: string;
-  identifier: string;
-  creator: string;
-  date: string;
-  downloadCount: number;
-  description: string;
-  detailsUrl: string;
-  downloadUrl: string;
-  xmlFileUrl: string;
-}
-
-export interface LevelsResponse {
-  count: number;
-  levels: Level[];
-}
+import { Level, LevelsResponse } from '../types';
 
 export async function fetchLevels(): Promise<LevelsResponse> {
   const { data, status, message } = await fetchServerData({
@@ -31,3 +15,5 @@ export async function fetchLevels(): Promise<LevelsResponse> {
     levels: data.levels as Level[],
   };
 }
+\nexport { Level, LevelsResponse } from '../types';
+
