+++ src/api/fetchServerEndpoints.ts
@@ -1,14 +1,11 @@
 import { stat } from 'fs/promises';
 import { promises as fs } from 'fs';
 import { getDirectories } from '../functions/fetchDirectories';
+import { Endpoint } from '../types';
 
 const SERVER_BASE_URL =
   typeof process !== 'undefined' && process.env?.SERVER_BASE_URL ? process.env.SERVER_BASE_URL : 'https://manic-launcher.vercel.app';
 
-interface Endpoint {
-  name: string;
-  endpoint: string;
-}
 
 export async function fetchServerEndpoints({
   routeName,
