+++ src/types/launcherSettings.ts
@@ -1,5 +1,1 @@
-export interface LauncherSettings {
-  playSoundOnInstall: boolean;
-  autoLaunchAfterInstall: boolean;
-  darkMode: boolean;
-}
+export { LauncherSettings } from './index';
